Fixes logging error in ConsulCache
If ConsulCache loses a connection to the agent then it will get an onFailure notification. That handler logs the error and then attempts to reconnect. However, the log line is malformed as a logger.error(...) cannot accept both string arguments and a throwable. This commit fixes that by putting the message + args in a String.format(...). This allows us to call LOGGER.error(String, Throwable).