basic support for the Consul ACL (providing token) + support for Basic Authentication (providing username+password)
Added support for the Consul ACL : a token can be provided to the Consul.Builder, this token will be added to every request as a query parameter.

Added support for Basic Authentication : in case you setup a reverse proxy in front of the Consul service (for additional security to the ACL, or as an alternative to the Consul ACL)

In both case, an Interceptor is added to the Retrofit client.