Override default backoff delay for Cache
Default backoff delay for caches is equal to 10s.
Enable this value to be overriden from properties if such a property exists.
For instance, "-Dcom.orbitz.consul.cache.backOffDelay=500".

I chose to use properties as the code of caches starts to grow fast, some
caches already have 4 constructors and i think it could be better to avoid
the creation of another one.