Override default backoff delay for Cache (#184)
* [Fix] KVCache does not work for exact key search

Commit 3714976 changes the way the prefix are removed from the key.
It works fine for key/subkey pattern but fails when we are looking for the
exact key (for intance, "a/b").

* Allows to override key extractor (#172)

* Add single line of documentation to example 4 (#176)

* Fix indentation and renumber examples

In example 4 there was an extra level of indentation, and there was no example 5, but examples 6, and 7, 8. I fixed the indentation and the numbering gap.

* add slight documentation to example 4

I added a single line of documentation to example 4 as per #174.

* Revert "Fix indentation and renumber examples"

* Fix indentation and renumber examples (#175)

In example 4 there was an extra level of indentation, and there was no example 5, but examples 6, and 7, 8. I fixed the indentation and the numbering gap.

* allow 404s on kv grabs

* fixing KV tests for 404

* 0.13.2

* Add support for Check-And-Set option to DELETE operation (#178)

* Add eclipse project files to .gitignore

* Add support for Check-And-Set option to DELETE operation.

* transaction support (#181)

* operator api support (#182)

* Override default backoff delay for Cache

Default backoff delay for caches is equal to 10s.
Enable this value to be overriden from properties if such a property exists.
For instance, "-Dcom.orbitz.consul.cache.backOffDelay=500".

I chose to use properties as the code of caches starts to grow fast, some
caches already have 4 constructors and i think it could be better to avoid
the creation of another one.