Clarify comment about having to continuously check in for TTL checks
If someone is only starting out with Consul they may think calling `pass()` once, like in the example, is enough. For example they may think the client library will ensure the Dead Man Switch is maintained. Moving the comment about the user being responsible to be next to the `pass()` call should make it more clear for new users.