Shutdownable ExecutorService (#222)
* Introduced a destroy() method on the Consul Object, so that the ThreadPool instanciated internally can be destroyed.

Make the ExecutorService a field of the Consul Object, so that it can be shutdown when the destroy method is invoked.

* Introduced a method to inject (into the Consul client's Builder) an instance of ExecutorService, that can be created, customized and managed by the application.

In case the ExecutorService needs to be customized and/or managed by the application, it can be injected into the Builder. The application can either call the Consul#destroy() method, or directly shutdown the ExecutorService.