Allow injecting Config
Currently, it is only possible to use default configuration provided by typesafe.
If the user wants to do something more specific, it must provide its configuration.
Allow it by injecting the Config in the Consul builder.

BTW, it also means that we can create several Consul instance with different configuration.