Upgrade Consul version in test to 1.0.3 (#304)
In this version, we need to:
 - enabled script checks in command line
This is needed to run test AgentTests#shouldRegisterNodeScriptCheck
According to changelog: 0.9.0 (July 20, 2017) "agent: Added a new enable_script_checks configuration option that defaults to false, meaning that in order to allow an agent to run health checks that execute scripts, this will need to be configured and set to true. This provides a safer out-of-the-box configuration for Consul where operators must opt-in to allow script-based health checks. [GH-3087]"
 - add "agent -dev -client 0.0.0.0" in docker command line:
This is the default command run by the docker image. However, we cannot add option without adding the default command line in the actual command line
 - set true to letters and numbers in RandomStringUtils#random
This is needed to ensure that the random string contains only numbers and letters i.e. printable characters.
According to changelog: 1.0.3 (Junuary 24, 2017) "agent: Updated Consul's HTTP server to ban all URLs containing non-printable characters (a bad request status will be returned for these cases). This affects some user-facing areas like key/value entry key names which are carried in URLs. [GH-3762]"