Enable custom log function in case of cache refresh error
We need to log sometimes as error, sometimes as warning.
Also, we would like to not log the same error every time (wait a few minutes in between).
This is specific to our use case.

Hence, we here allow setting a custom consumer to log on cache refresh error.
The behavior of caches for those who don't need this feature will remain unchanged.