[Fix] Agent/force-leave does not work
This was due to the fact that the `node` parameter was missing.
The endpoint agent/force-leave does not exist and Consul respond with 301.
The request was then redirected to agent/force-leave/ by OkHttp client.
This redirection is always done with GET and Consul rejects it once more.

The API documentation of Consul HTTP API (hashicorp) is out-of-date.
To see the actual endpoint see:
 - https://github.com/hashicorp/consul/blob/master/agent/http_oss.go#L25
 - https://github.com/hashicorp/consul/blob/master/agent/agent_endpoint.go#L323
 - https://github.com/hashicorp/consul/blob/master/agent/agent.go#L1442