A 400 series error should not blacklist the consul instance that is (#362)
responding. This is because a 404 or 403 error is actually completely
valid.

This does remove some potential functionality (such as being able to do
cross-cluster failover), however that functionality is not exactly
supported anyways since it introduces the potential to set a key in one
cluster and then fail to watch it in another.