Allow to specify cache control as well as getting information about cache HIT/MISS
While the first patch was working, it was a bit insufficient as it did not allow
to handle precise cache Control as suggested by @Aestek

I took the path to change the enum ConsistencyMode into a class emulating enum's
methods in order to keep backwards compatibility and avoiding breaking clients.

This patch is a bit intrusive since I have to dynamically change Header Cache-Control
during requests and current static ways to inject headers using injectors was not
sufficient. The mechanism might be extended later to dynamically inject other kind of
headers if necessary.

I also added a way to get Cache result information when available in Consul Response.

Cache has been implemented only for endpoints supporting it.